[internal/driver/glfw] move non-darwin type and function into appropriate file
This fixes `golangci-lint run ./... --enable-only unused`
